Move typechain contracts to src/contracts
